audio_tagger/requirements.txt
numpy<2.0
tensorflow-cpu==2.15.0
numba==0.58.1
soundfile              
librosa
boto3